Fixed errors from object to array, and renamed task module to tasks.
