+++ .gitignore
@@ -22,3 +22,5 @@ yarn-error.log*
 
 #Electron-builder output
 /dist_electron
+
+.envrc
