+++ src/store/index.js
@@ -3,14 +3,14 @@ import Vuex from 'vuex'
 import createPersistedState from 'vuex-persistedstate'
 
 import device from '@/store/modules/device'
-import task from '@/store/modules/task'
+import tasks from '@/store/modules/tasks'
 
 Vue.use(Vuex)
 
 export default new Vuex.Store({
   modules: {
     device,
-    task
+    tasks
   },
   plugins: [createPersistedState()]
 })
