+++ src/store/modules/device/getters.js
@@ -1,0 +1,4 @@
+export default {
+  deviceToken: ({ deviceToken }) => deviceToken,
+  deviceUuid: ({ deviceUuid }) => deviceUuid
+}
