+++ src/store/modules/tasks/getters.js
@@ -1,0 +1,5 @@
+export default {
+  errors: ({ errors }) => errors,
+  focusTargetId: ({ focusTargetId }) => focusTargetId,
+  tasks: ({ tasks }) => tasks
+}
