+++ src/store/modules/tasks/mutations.js
@@ -8,10 +8,13 @@ export default {
     state.isSubmitting = false
     state.isInitialized = true
     state.tasks = tasks
+    state.error = []
   },
-  [types.FAILURE_FETCH_TASKS] (state, errors = {}) {
-    state.errors = errors
+  [types.FAILURE_FETCH_TASKS] (state, error) {
     state.isSubmitting = false
+    if (error) {
+      state.errors.push(error)
+    }
   },
   [types.REQUEST_FETCH_TASK] (state) {
     state.isSubmitting = true
@@ -20,10 +23,13 @@ export default {
     state.isSubmitting = false
     state.isInitialized = true
     state.tasks = [task]
+    state.error = []
   },
-  [types.FAILURE_FETCH_TASK] (state, errors = {}) {
-    state.errors = errors
+  [types.FAILURE_FETCH_TASK] (state, error) {
     state.isSubmitting = false
+    if (error) {
+      state.errors.push(error)
+    }
   },
   [types.SET_FOCUS_TARGET_ID] (state, { id }) {
     state.focusTargetId = id
